requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
unidecode==1.3.7
